refactor: Move OpenAPI-specifi `endpoints` implementation to the `BaseOpenAPISchema` class
