+++ src/schemathesis/specs/openapi/schemas.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, Generator, Iterator, List, Optional, Tuple
 from urllib.parse import urljoin, urlsplit
 
 import jsonschema
+from requests.structures import CaseInsensitiveDict
 
 from ...exceptions import InvalidSchema
 from ...filters import should_skip_by_operation_id, should_skip_by_tag, should_skip_endpoint, should_skip_method
@@ -27,6 +28,14 @@ class BaseOpenAPISchema(BaseSchema):
     def __repr__(self) -> str:
         info = self.raw_schema["info"]
         return f"{self.__class__.__name__} for {info['title']} ({info['version']})"
+
+    @property
+    def endpoints(self) -> Dict[str, CaseInsensitiveDict]:
+        if not hasattr(self, "_endpoints"):
+            # pylint: disable=attribute-defined-outside-init
+            endpoints = self.get_all_endpoints()
+            self._endpoints = endpoints_to_dict(endpoints)
+        return self._endpoints
 
     @property
     def resolver(self) -> ConvertingResolver:
@@ -310,3 +319,11 @@ def get_common_parameters(methods: Dict[str, Any]) -> List[Dict[str, Any]]:
     if common_parameters is not None:
         return deepcopy(common_parameters)
     return []
+
+
+def endpoints_to_dict(endpoints: Generator[Endpoint, None, None]) -> Dict[str, CaseInsensitiveDict]:
+    output: Dict[str, CaseInsensitiveDict] = {}
+    for endpoint in endpoints:
+        output.setdefault(endpoint.path, CaseInsensitiveDict())
+        output[endpoint.path][endpoint.method] = endpoint
+    return output
