+++ test/test_direct_access.py
@@ -4,7 +4,7 @@ from requests.structures import CaseInsensitiveDict
 
 import schemathesis
 from schemathesis.models import Case, Endpoint
-from schemathesis.schemas import endpoints_to_dict
+from schemathesis.specs.openapi.schemas import endpoints_to_dict
 
 
 def test_contains(swagger_20):
@@ -13,7 +13,7 @@ def test_contains(swagger_20):
 
 def test_getitem(simple_schema, mocker):
     swagger = schemathesis.from_dict(simple_schema)
-    mocked = mocker.patch("schemathesis.schemas.endpoints_to_dict", wraps=endpoints_to_dict)
+    mocked = mocker.patch("schemathesis.specs.openapi.schemas.endpoints_to_dict", wraps=endpoints_to_dict)
     assert "_endpoints" not in swagger.__dict__
     assert isinstance(swagger["/v1/users"], CaseInsensitiveDict)
     assert mocked.call_count == 1
